Add files via upload
This game is developed in IntelliJ IDE using java as its programming language. Flappy Bird is an endless game involving a bird the player can control. The player has to save the bird from colliding with the hurdles like pipes. Each time the bird passes through the pipes, the score gets incremented by one. The game ends when the bird crashes with the pipes or falls down due to gravity.